ZAPP-364 Move towards consolidating the edit commit flow for attribute and attribute reporting into one file
