+++ src/components/ZclAttributeReportingManager.vue
@@ -134,7 +134,7 @@ limitations under the License.
                   ? 'visible'
                   : 'hidden',
               }"
-              @click="resetAttribute(props.row.id)"
+              @click="resetAttributeReporting(props.row.id)"
             />
             <q-btn
               dense
@@ -143,8 +143,14 @@ limitations under the License.
               color="blue"
               @click="
                 editableAttributes[props.row.id]
-                  ? commitEdittedAttribute(props.row, selectedCluster.id)
-                  : setEditableAttribute(props.row.id, selectedCluster.id)
+                  ? commitEdittedAttributeReporting(
+                      props.row,
+                      selectedCluster.id
+                    )
+                  : setEditableAttributeReporting(
+                      props.row.id,
+                      selectedCluster.id
+                    )
               "
             />
           </q-td>
@@ -155,159 +161,12 @@ limitations under the License.
 </template>
 
 <script>
-import * as Util from '../util/util'
-import Vue from 'vue'
+//This mixin derives from common-mixin.
+import EditableAttributeMixin from '../util/editable-attributes-mixin'
 
-import CommonMixin from '../util/common-mixin'
 export default {
   name: 'ZclAttributeReportingManager',
-  mixins: [CommonMixin],
-  methods: {
-    handleLocalSelection(list, attributeDataId, clusterId) {
-      let hash = this.hashAttributeIdClusterId(attributeDataId, clusterId)
-      let indexOfValue = list.indexOf(hash)
-      if (indexOfValue === -1) {
-        list.push(hash)
-      } else {
-        list.splice(indexOfValue, 1)
-      }
-    },
-    handleLocalChange(value, list, hash) {
-      Vue.set(this[list], hash, value)
-      this[list] = Object.assign({}, this[list])
-    },
-    toggleAttributeSelection(list, listType, attributeData, clusterId, enable) {
-      // We determine the ID that we need to toggle within the list.
-      // This ID comes from hashing the base ZCL attribute and cluster data.
-      let indexOfValue = list.indexOf(
-        this.hashAttributeIdClusterId(attributeData.id, clusterId)
-      )
-      let addedValue
-      if (indexOfValue === -1) {
-        addedValue = true
-      } else {
-        addedValue = false
-      }
-
-      let editContext = {
-        action: 'boolean',
-        endpointTypeId: this.selectedEndpointTypeId,
-        id: attributeData.id,
-        value: addedValue,
-        listType: listType,
-        clusterRef: clusterId,
-        attributeSide: attributeData.side,
-      }
-      this.$store.dispatch('zap/updateSelectedAttribute', editContext)
-    },
-    setAttributeSelection(listType, attributeData, clusterId, enable) {
-      let editContext = {
-        action: 'boolean',
-        endpointTypeId: this.selectedEndpointTypeId,
-        id: attributeData.id,
-        value: enable,
-        listType: listType,
-        clusterRef: clusterId,
-        attributeSide: attributeData.side,
-      }
-      this.$store.dispatch('zap/updateSelectedAttribute', editContext)
-    },
-    handleAttributeDefaultChange(newValue, listType, attributeData, clusterId) {
-      let editContext = {
-        action: 'text',
-        endpointTypeId: this.selectedEndpointTypeId,
-        id: attributeData.id,
-        value: newValue,
-        listType: listType,
-        clusterRef: clusterId,
-        attributeSide: attributeData.side,
-      }
-      this.$store.dispatch('zap/updateSelectedAttribute', editContext)
-    },
-    isDefaultValueValid(id) {
-      return this.defaultValueValidation[id] != null
-        ? this.defaultValueValidation[id].length === 0
-        : true
-    },
-    getDefaultValueErrorMessage(id) {
-      return this.defaultValueValidation[id] != null
-        ? this.defaultValueValidation[id].reduce(
-            (validationIssueString, currentVal) => {
-              return validationIssueString + '\n' + currentVal
-            },
-            ''
-          )
-        : ''
-    },
-    hashAttributeIdClusterId(attributeId, clusterId) {
-      return Util.cantorPair(attributeId, clusterId)
-    },
-
-    initializeTextEditableList(originatingList, editableList, attrClusterHash) {
-      let data = originatingList[attrClusterHash]
-      editableList[attrClusterHash] = data
-    },
-
-    setEditableAttribute(attributeId, selectedClusterId) {
-      let attrClusterHash = this.hashAttributeIdClusterId(
-        attributeId,
-        selectedClusterId
-      )
-
-      this.initializeTextEditableList(
-        this.selectionMin,
-        this.editableMin,
-        attrClusterHash
-      )
-
-      this.initializeTextEditableList(
-        this.selectionMax,
-        this.editableMax,
-        attrClusterHash
-      )
-
-      this.initializeTextEditableList(
-        this.selectionReportableChange,
-        this.editableReportable,
-        attrClusterHash
-      )
-
-      this.$store.dispatch('zap/setAttributeEditting', {
-        attributeId: attributeId,
-        editState: true,
-      })
-    },
-
-    resetAttribute(attributeId) {
-      this.$store.dispatch('zap/setAttributeEditting', {
-        attributeId: attributeId,
-        editState: false,
-      })
-    },
-
-    commitEdittedAttribute(attributeData, clusterId) {
-      let hash = this.hashAttributeIdClusterId(attributeData.id, clusterId)
-
-      this.handleAttributeDefaultChange(
-        this.editableMin[hash],
-        'reportingMin',
-        attributeData,
-        clusterId
-      )
-      this.handleAttributeDefaultChange(
-        this.editableMax[hash],
-        'reportingMax',
-        attributeData,
-        clusterId
-      )
-
-      this.$store.dispatch('zap/setAttributeEditting', {
-        attributeId: attributeData.id,
-        editState: false,
-      })
-    },
-  },
-
+  mixins: [EditableAttributeMixin],
   computed: {
     attributeData: {
       get() {
@@ -361,15 +220,12 @@ export default {
     },
     editableAttributes: {
       get() {
-        return this.$store.state.zap.attributeView.editableAttributes
+        return this.$store.state.zap.attributeView.editableAttributesReporting
       },
     },
   },
   data() {
     return {
-      editableMin: {},
-      editableMax: {},
-      editableReportable: {},
       pagination: {
         rowsPerPage: 0,
       },
