+++ src/store/zap/state.js
@@ -94,6 +94,7 @@ export default function () {
     attributeView: {
       selectedAttributes: [],
       editableAttributes: {},
+      editableAttributesReporting: {},
       selectedSingleton: [],
       selectedBounded: [],
       defaultValue: {},
